+++ internal/gohacks/stringss/escape_newlines.go
@@ -1,0 +1,7 @@
+package stringss
+
+import "strings"
+
+func EscapeNewLines(text string) string {
+	return strings.ReplaceAll(text, "\n", "\\n")
+}
