--- internal/gohacks/stringss/strings_test.go
@@ -1,65 +1,0 @@
-package stringss_test
-
-import (
-	"testing"
-
-	"github.com/git-town/git-town/v23/internal/gohacks/stringss"
-	"github.com/shoenig/test/must"
-)
-
-func TestEscapeNewLines(t *testing.T) {
-	t.Parallel()
-	tests := map[string]string{
-		"":                    "",
-		"no newlines":         "no newlines",
-		"one\nnewline":        "one\\nnewline",
-		"two\nnew\nlines":     "two\\nnew\\nlines",
-		"three\nnew\nlines\n": "three\\nnew\\nlines\\n",
-	}
-	for give, want := range tests {
-		have := stringss.EscapeNewLines(give)
-		must.EqOp(t, want, have)
-	}
-}
-
-func TestIndent(t *testing.T) {
-	t.Parallel()
-
-	t.Run("empty", func(t *testing.T) {
-		t.Parallel()
-		have := stringss.IndentLines("", 2)
-		must.EqOp(t, "", have)
-	})
-
-	t.Run("indent by 2 spaces", func(t *testing.T) {
-		t.Parallel()
-		give := "one\ntwo\nthree"
-		have := stringss.IndentLines(give, 2)
-		want := "  one\n  two\n  three"
-		must.EqOp(t, want, have)
-	})
-
-	t.Run("indent by 4 spaces", func(t *testing.T) {
-		t.Parallel()
-		give := "one\ntwo\nthree"
-		have := stringss.IndentLines(give, 4)
-		want := "    one\n    two\n    three"
-		must.EqOp(t, want, have)
-	})
-}
-
-func TestLeadingWhitespace(t *testing.T) {
-	t.Parallel()
-	tests := map[string]string{
-		"    text":   "    ",   // spaces
-		"\t\ttext":   "\t\t",   // tabs
-		"  \t  text": "  \t  ", // mixed spaces and tabs
-		"text":       "",       // no indentation
-		"":           "",       // empty string
-		"    ":       "    ",   // only whitespace
-	}
-	for give, want := range tests {
-		have := stringss.LeadingWhitespace(give)
-		must.EqOp(t, want, have)
-	}
-}
