+++ internal/gohacks/stringss/zero_delineated.go
